Time: 0 ms (100%), Space: 8.9 MB (8.1%) - LeetHub

+++ 0050-powx-n/0050-powx-n.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     double myPow(double x, int n) {
-        long double ans=pow(x,n);
+        double ans=pow(x,n);
         return ans;
     }
 };
